:recycle: Refactor: 엔티티 도메인에 맞춰 이동 (#52)
