+++ src/main/java/baeksaitong/sofp/domain/favorite/service/FavoriteService.java
@@ -9,9 +9,9 @@ import baeksaitong.sofp.domain.favorite.repository.FavoriteRepository;
 import baeksaitong.sofp.domain.pill.error.PillErrorCode;
 import baeksaitong.sofp.domain.pill.repository.PillRepository;
 import baeksaitong.sofp.domain.profile.service.ProfileService;
-import baeksaitong.sofp.global.common.entity.Favorite;
-import baeksaitong.sofp.global.common.entity.Pill;
-import baeksaitong.sofp.global.common.entity.Profile;
+import baeksaitong.sofp.domain.favorite.entity.Favorite;
+import baeksaitong.sofp.domain.pill.entity.Pill;
+import baeksaitong.sofp.domain.profile.entity.Profile;
 import baeksaitong.sofp.global.error.exception.BusinessException;
 import baeksaitong.sofp.global.s3.AwsS3Service;
 import lombok.RequiredArgsConstructor;
