+++ src/main/java/baeksaitong/sofp/domain/health/service/DiseaseAllergyService.java
@@ -5,9 +5,9 @@ import baeksaitong.sofp.domain.health.dto.response.DiseaseAllergyRes;
 import baeksaitong.sofp.domain.health.repository.DiseaseAllergyRepository;
 import baeksaitong.sofp.domain.health.repository.ProfileDiseaseAllergyRepository;
 import baeksaitong.sofp.domain.profile.service.ProfileService;
-import baeksaitong.sofp.global.common.entity.DiseaseAllergy;
-import baeksaitong.sofp.global.common.entity.Profile;
-import baeksaitong.sofp.global.common.entity.ProfileDiseaseAllergy;
+import baeksaitong.sofp.domain.health.entity.DiseaseAllergy;
+import baeksaitong.sofp.domain.profile.entity.Profile;
+import baeksaitong.sofp.domain.health.entity.ProfileDiseaseAllergy;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
