+++ src/main/java/baeksaitong/sofp/domain/history/service/HistoryService.java
@@ -8,9 +8,9 @@ import baeksaitong.sofp.domain.history.error.HistoryErrorCode;
 import baeksaitong.sofp.domain.history.repository.HistoryRepository;
 import baeksaitong.sofp.domain.pill.repository.PillRepository;
 import baeksaitong.sofp.domain.profile.service.ProfileService;
-import baeksaitong.sofp.global.common.collection.History;
-import baeksaitong.sofp.global.common.entity.Pill;
-import baeksaitong.sofp.global.common.entity.Profile;
+import baeksaitong.sofp.domain.history.collection.History;
+import baeksaitong.sofp.domain.pill.entity.Pill;
+import baeksaitong.sofp.domain.profile.entity.Profile;
 import baeksaitong.sofp.global.error.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
