+++ src/main/java/baeksaitong/sofp/domain/member/entity/enums/MemberRole.java
@@ -1,4 +1,4 @@
-package baeksaitong.sofp.global.common.entity.enums;
+package baeksaitong.sofp.domain.member.entity.enums;
 
 public enum MemberRole {
     ROLE_USER, ROLE_ADMIN
