+++ src/main/java/baeksaitong/sofp/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package baeksaitong.sofp.domain.member.repository;
 
-import baeksaitong.sofp.global.common.entity.Member;
+import baeksaitong.sofp.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
