+++ src/main/java/baeksaitong/sofp/domain/pill/service/PillService.java
@@ -6,9 +6,9 @@ import baeksaitong.sofp.domain.pill.dto.response.PillRes;
 import baeksaitong.sofp.domain.pill.repository.PillRepository;
 import baeksaitong.sofp.domain.pill.repository.ProfilePillRepository;
 import baeksaitong.sofp.domain.profile.service.ProfileService;
-import baeksaitong.sofp.global.common.entity.Pill;
-import baeksaitong.sofp.global.common.entity.Profile;
-import baeksaitong.sofp.global.common.entity.ProfilePill;
+import baeksaitong.sofp.domain.pill.entity.Pill;
+import baeksaitong.sofp.domain.profile.entity.Profile;
+import baeksaitong.sofp.domain.pill.entity.ProfilePill;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
