+++ src/main/java/baeksaitong/sofp/domain/profile/dto/response/ProfileBasicRes.java
@@ -1,6 +1,6 @@
 package baeksaitong.sofp.domain.profile.dto.response;
 
-import baeksaitong.sofp.global.common.entity.Profile;
+import baeksaitong.sofp.domain.profile.entity.Profile;
 
 public record ProfileBasicRes(Long id, String name, String imgURL, String color) {
     public ProfileBasicRes(Profile profile){
