+++ src/main/java/baeksaitong/sofp/domain/profile/entity/enums/Gender.java
@@ -1,0 +1,12 @@
+package baeksaitong.sofp.domain.profile.entity.enums;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+
+public enum Gender {
+    MALE, FEMALE;
+
+    @JsonCreator
+    public static Gender from(String s) {
+        return Gender.valueOf(s.toUpperCase());
+    }
+}
