+++ src/main/java/baeksaitong/sofp/domain/profile/repository/ProfileRepository.java
@@ -1,7 +1,7 @@
 package baeksaitong.sofp.domain.profile.repository;
 
-import baeksaitong.sofp.global.common.entity.Member;
-import baeksaitong.sofp.global.common.entity.Profile;
+import baeksaitong.sofp.domain.member.entity.Member;
+import baeksaitong.sofp.domain.profile.entity.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,6 +11,5 @@ import java.util.Optional;
 @Repository
 public interface ProfileRepository extends JpaRepository<Profile, Long> {
     Long countByMember(Member member);
-    Optional<Profile> findByNameAndMember(String name, Member member);
     List<Profile> findAllByMember(Member member);
 }
