+++ src/main/java/baeksaitong/sofp/domain/profile/service/ProfileService.java
@@ -5,8 +5,8 @@ import baeksaitong.sofp.domain.profile.dto.response.ProfileBasicRes;
 import baeksaitong.sofp.domain.profile.dto.response.ProfileDetailRes;
 import baeksaitong.sofp.domain.profile.error.ProfileErrorCode;
 import baeksaitong.sofp.domain.profile.repository.ProfileRepository;
-import baeksaitong.sofp.global.common.entity.Member;
-import baeksaitong.sofp.global.common.entity.Profile;
+import baeksaitong.sofp.domain.member.entity.Member;
+import baeksaitong.sofp.domain.profile.entity.Profile;
 import baeksaitong.sofp.global.error.exception.BusinessException;
 import baeksaitong.sofp.global.redis.RedisPrefix;
 import baeksaitong.sofp.global.redis.RedisService;
