+++ src/main/java/baeksaitong/sofp/domain/search/repository/PillRepositoryCustom.java
@@ -1,7 +1,7 @@
 package baeksaitong.sofp.domain.search.repository;
 
 import baeksaitong.sofp.domain.search.dto.request.KeywordReq;
-import baeksaitong.sofp.global.common.entity.Pill;
+import baeksaitong.sofp.domain.pill.entity.Pill;
 import org.springframework.data.domain.Page;
