--- src/main/java/baeksaitong/sofp/global/common/entity/enums/MemberGender.java
@@ -1,12 +1,0 @@
-package baeksaitong.sofp.global.common.entity.enums;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-
-public enum MemberGender {
-    MALE, FEMALE;
-
-    @JsonCreator
-    public static MemberGender from(String s) {
-        return MemberGender.valueOf(s.toUpperCase());
-    }
-}
